Overhaul the fonts
Remove unused fonts, create a list of licenses (Licenses.md), and add
license files where available.

In the case of Thundercats, the license specifically forbids redistribution,
so I’ve written to the copyright holder to ask permission.

In the case of Whiteboard, I cannot find definitive evidence of a license,
so I’ve written to the creator to ask.